# Rewrite and benchmark XML output for IR