--- a/benchmark/xmlBenchmark.ts
+++ b/benchmark/xmlBenchmark.ts
@@ -1,12 +1,23 @@
 /* tslint:disable:no-console */
 import { Event, Suite } from "benchmark";
+import { IRBlockHandler } from "../src/ir/IRBlockHandler";
 import { toXML } from "../src/output/xml";
-import { parse } from "../src/parser/BlockParser";
+import { BlockParser } from "../src/parser/BlockParser";
+import { SchemaDecorator } from "../src/schema/Schema";
+import { Cardinality, INVALID_TAG, ROOT, SchemaDefinition } from "../src/schema/SchemaDefinition";
 import { formatBytes, generateBenchmarkInput } from "./utils";
 
 const input = generateBenchmarkInput();
 console.log("Size of test input: " + formatBytes(input.length));
-const ast = parse(input);
+
+const parser = new BlockParser(
+	new IRBlockHandler(new SchemaDecorator(getAllowAllSchema()), item => item)
+);
+
+const ast = parser.parse(input);
+if (ast === null) {
+	throw new Error("AST was null");
+}
 
 new Suite("XML serialization")
 	.add("toXML", () => {
@@ -24,3 +35,28 @@ new Suite("XML serialization")
 	})
 	// run async
 	.run({ async: true });
+
+function getAllowAllSchema(): SchemaDefinition {
+	const blockProps = [
+		{ name: "children", content: [{ tag: INVALID_TAG, cardinality: Cardinality.ZeroOrMore }] }
+	];
+
+	const blockContent = {
+		head: { name: "head", content: [] },
+		props: blockProps,
+		defaultTag: "_default"
+	};
+
+	return {
+		blocks: {
+			[ROOT]: blockContent,
+			["_default"]: blockContent,
+			[INVALID_TAG]: blockContent
+		},
+		inline: {
+			[INVALID_TAG]: {
+				props: [{ name: "arg", content: [{ tag: INVALID_TAG, schema: "[base]" }] }]
+			}
+		}
+	};
+}