--- a/test/_resources/output/xml/default_root.xml
+++ b/test/_resources/output/xml/default_root.xml
@@ -1,5 +1,9 @@
 <root>
-	<_default>
-		<head>Hello World</head>
-	</_default>
+	<content>
+		<paragraph>
+			<text>
+				Hello World
+			</text>
+		</paragraph>
+	</content>
 </root>