--- a/test/_resources/output/xml/empty.xml
+++ b/test/_resources/output/xml/empty.xml
@@ -1,1 +1,3 @@
-<root/>
+<root>
+	<content/>
+</root>