--- a/test/_resources/output/xml/escape.xml
+++ b/test/_resources/output/xml/escape.xml
@@ -1,14 +1,24 @@
 <root>
-	<_default>
-		<head>Lorem # Ipsum</head>
-	</_default>
-	<_default>
-		<head># Lorem Ipsum</head>
-	</_default>
-	<_default>
-		<head>Lorem Ispum #</head>
-	</_default>
-	<_default>
-		<head>\ # [ ]</head>
-	</_default>
+	<content>
+		<paragraph>
+			<text>
+				Lorem # Ipsum
+			</text>
+		</paragraph>
+		<paragraph>
+			<text>
+				# Lorem Ipsum
+			</text>
+		</paragraph>
+		<paragraph>
+			<text>
+				Lorem Ispum #
+			</text>
+		</paragraph>
+		<paragraph>
+			<text>
+				\ # [ ]
+			</text>
+		</paragraph>
+	</content>
 </root>