--- a/test/_resources/output/xml/inline_simple.xml
+++ b/test/_resources/output/xml/inline_simple.xml
@@ -1,5 +1,13 @@
 <root>
-	<_default>
-		<head><strong>id</strong></head>
-	</_default>
+	<content>
+		<paragraph>
+			<text>
+				<strong>
+					<text>
+						id
+					</text>
+				</strong>
+			</text>
+		</paragraph>
+	</content>
 </root>