This PR rewrites `toXML` for the IR. I added tests and a performance benchmark.

The previous version of `toXML` added indentation recursively (a single indent at a time), while the current one only indents them once by the correct at the very end. Surprisingly, a little performance is lost, and I don't really know why :cry: 

<details><summary>Benchmark results</summary>
Before:

```
Size of test input: 7.44 MB
toXML x 3.05 ops/sec ±9.48% (13 runs sampled)
JSON.stringify x 4.00 ops/sec ±9.25% (14 runs sampled)
JSON/XML ratio: 0.76
```

After:

```
Size of test input: 7.44 MB
toXML x 2.55 ops/sec ±4.87% (11 runs sampled)
JSON.stringify x 4.78 ops/sec ±3.21% (16 runs sampled)
JSON/XML ratio: 0.53
```

</details>